Add `styles.css` to copied files
The [plugin generator](https://github.com/mnaoumov/generator-obsidian-plugin) creates a `styles.css` file on the top level, so my assumption is the intent is for that file to be included in the files which get copied to the `dist` directory. It's fairly straightforward to treat it similarly to `manifest.json`, but I don't use sass so am uncertain if it will play nicely with the logic building from that.